agent/requirements.txt
openai
requests
python-dotenv
flask